pyproject.toml
[project]
name = "tara-migrate"
version = "1.0.0"
description = "TARA Shopify store migration tool"
requires-python = ">=3.9"
dependencies = [
    "requests>=2.31.0",
    "openai>=1.50.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ruff>=0.4.0",
]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["F841", "E741", "E501"]
